Classification SVM model of ozone Machine Learning
I developed 2 machine learning software that predict and classify ozone day and non-ozone day. The working principle of the two is similar but there are differences. I got the dataset from ics.icu. Each software has a different mathematical model, Gaussian RBF and Linear Kernel, and classifications are visualized in different ways. I would be happy to present the software to you!